Implemented GitHub API and added 404 page
Retrieves project data from GitHub using their API. Also, I installed marked to convert the markdown syntax github uses to html. I did this so I could grab the readme from a github project repo and display it on my website. Without this library, the readme would be displayed as a messy wall of text. Lastly, I added a 404 redirect so that if someone enters the url for a non existant page, they will be redirected to a 404 page not found.